--- playground/sourcemaps/src/App.jsx
@@ -1,25 +1,0 @@
-import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
-import './App.css'
-import { lazy } from 'react'
-
-const SubApp = lazy(() => import('./SubApp'))
-
-function App() {
-  const [count, setCount] = useState(0)
-
-  return (
-    <>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button onClick={() => setCount((count) => count + 1)}>
-          count is {count}
-        </button>
-      </div>
-      <SubApp />
-    </>
-  )
-}
-
-export default App
