--- playground/sourcemaps/src/SubApp.jsx
@@ -1,9 +1,0 @@
-function SubApp() {
-  return (
-      <div>
-        This component should be lazily loaded
-      </div>
-  )
-}
-
-export default SubApp
