+++ playground/sourcemaps/src/SubApp.tsx
@@ -1,0 +1,5 @@
+function SubApp() {
+    return <div>This component should be lazily loaded</div>
+}
+
+export default SubApp
