+++ playground/sourcemaps/src/main.tsx
@@ -1,0 +1,13 @@
+import { StrictMode } from 'react'
+import { createRoot } from 'react-dom/client'
+import App from './App.js'
+
+const root = document.getElementById('root')
+
+if (root) {
+    createRoot(root).render(
+        <StrictMode>
+            <App />
+        </StrictMode>
+    )
+}
